# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Crypto Trader API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: Get real order history across linked broker accounts
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
        "401":
          description: Unauthorized
    post:
      operationId: placeTrade
      tags: [trades]
      summary: Place a real-money market order, routed to the right broker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Invalid order or broker error
        "401":
          description: Unauthorized
...
  /auto-trade/settings:
    get:
      operationId: getAutoTradeSettings
      tags: [auto-trade]
      summary: Get auto-trade settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoTradeSettings"
    put:
      operationId: updateAutoTradeSettings
      tags: [auto-trade]
      summary: Update auto-trade settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoTradeSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoTradeSettings"
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerAuthUser
      summary: Create an email/password account and session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "201":
          description: Account created and session established.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "400":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Email already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/login:
    post:
      tags: [Auth]
      operationId: loginAuthUser
      summary: Authenticate an email/password account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Session established.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Invalid email or password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      operationId: requestPasswordReset
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Generic response that does not reveal whether the email exists.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /account/ai-credentials:
    get:
      tags: [account]
      operationId: getAiCredentials
      summary: Get the current user's configured AI providers
      responses:
        "200":
          description: Provider configuration status without raw credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiCredentialsStatus"
        "401":
          description: Unauthorized
...
  /broker/connect:
    post:
      tags: [broker]
      operationId: connectBroker
      summary: Link an Alpaca brokerage account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectBrokerInput"
      responses:
        "200":
          description: Broker connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerStatus"
        "400":
          description: Invalid credentials
        "401":
          description: Unauthorized
...
  /broker/orders:
    get:
      tags: [broker]
      operationId: listBrokerOrders
      summary: List recent orders in the linked brokerage account
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BrokerOrder"
        "400":
          description: No broker connected
        "401":
          description: Unauthorized
    post:
      tags: [broker]
      operationId: placeBrokerOrder
      summary: Place a real-money market order (notional, fractional supported)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBrokerOrderInput"
      responses:
        "200":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerOrder"
        "400":
          description: Invalid order or broker error
        "401":
          description: Unauthorized
  /broker/auto-trade:
    get:
      tags: [broker]
      operationId: getBrokerAutoTradeSettings
      summary: Get real-money auto-trade settings for the linked broker account
      responses:
        "200":
          description: Real-money auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerAutoTradeSettings"
        "400":
          description: No broker connected
        "401":
          description: Unauthorized
    put:
      tags: [broker]
      operationId: updateBrokerAutoTradeSettings
      summary: Update real-money auto-trade settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrokerAutoTradeInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerAutoTradeSettings"
        "400":
          description: No broker connected or invalid input
        "401":
          description: Unauthorized
  /broker/oanda/connect:
    post:
      tags: [broker]
      operationId: connectOanda
      summary: Link an OANDA account for forex, commodity, and index CFD trading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectOandaInput"
      responses:
        "200":
          description: OANDA connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaStatus"
        "400":
          description: Invalid credentials
        "401":
          description: Unauthorized
...
  /broker/oanda/auto-trade:
    get:
      tags: [broker]
      operationId: getOandaAutoTradeSettings
      summary: Get real-money auto-trade settings for the linked OANDA account
      responses:
        "200":
          description: OANDA auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaAutoTradeSettings"
        "400":
          description: No OANDA account connected
        "401":
          description: Unauthorized
    put:
      tags: [broker]
      operationId: updateOandaAutoTradeSettings
      summary: Update real-money auto-trade settings for OANDA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrokerAutoTradeInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OandaAutoTradeSettings"
        "400":
          description: No OANDA account connected or invalid input
        "401":
          description: Unauthorized
  /broker/kraken/connect:
    post:
      tags: [broker]
      operationId: connectKraken
      summary: Link a Kraken Spot account for manual crypto trading
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectKrakenInput"
      responses:
        "200":
          description: Kraken connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KrakenStatus"
        "400":
          description: Invalid credentials
        "401":
          description: Unauthorized
...
  /broker/kraken/auto-trade:
    get:
      tags: [broker]
      operationId: getKrakenAutoTradeSettings
      summary: Get Kraken AI auto-trade limits
      responses:
        "200":
          description: Kraken AI auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KrakenAutoTradeSettings"
        "400":
          description: No Kraken account connected
        "401":
          description: Unauthorized
    put:
      tags: [broker]
      operationId: updateKrakenAutoTradeSettings
      summary: Update Kraken AI auto-trade settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBrokerAutoTradeInput"
      responses:
        "200":
          description: Updated Kraken AI auto-trade settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KrakenAutoTradeSettings"
        "400":
          description: No Kraken account connected or invalid settings
        "401":
          description: Unauthorized
components:
...
  schemas:
...
    AiCredentialStatus:
      type: object
      required: [provider, configured, maskedKey]
      properties:
        provider:
          type: string
          enum: [anthropic, openai, gemini]
        configured:
          type: boolean
        maskedKey:
          type: [string, "null"]
    AiCredentialsStatus:
      type: object
      required: [providers]
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/AiCredentialStatus"
...
    ConnectBrokerInput:
      type: object
      properties:
        apiKey:
          type: string
          minLength: 1
        apiSecret:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [paper, live]
      required: [apiKey, apiSecret, mode]
    BrokerStatus:
      type: object
      properties:
        connected:
          type: boolean
        mode:
          type: string
          enum: [paper, live]
        apiKeyMasked:
          type: string
        accountNumber:
          type: string
        currency:
          type: string
        cash:
          type: number
        buyingPower:
          type: number
        equity:
          type: number
        portfolioValue:
          type: number
        status:
          type: string
        longMarketValue:
          type: number
        shortMarketValue:
          type: number
        daytradeCount:
          type: number
        patternDayTrader:
          type: boolean
        shortingEnabled:
          type: boolean
        multiplier:
          type: number
        maintenanceMargin:
          type: number
        initialMargin:
          type: number
        lastEquity:
          type: number
        createdAt:
          type: string
      required: [connected]
...
    PlaceBrokerOrderInput:
      type: object
      properties:
        symbol:
          type: string
          minLength: 1
        side:
          type: string
          enum: [buy, sell]
        notional:
          type: number
          minimum: 1
      required: [symbol, side, notional]
    BrokerAutoTradeSettings:
      type: object
      properties:
        enabled:
          type: boolean
        maxPerTradeUsd:
          type: number
        dailyLimitUsd:
          type: number
        spentTodayUsd:
          type: number
        mode:
          type: string
          enum: [paper, live]
      required: [enabled, maxPerTradeUsd, dailyLimitUsd, spentTodayUsd, mode]
    UpdateBrokerAutoTradeInput:
      type: object
      properties:
        enabled:
          type: boolean
        maxPerTradeUsd:
          type: number
          minimum: 1
          maximum: 10000
        dailyLimitUsd:
          type: number
          minimum: 1
          maximum: 100000
    ConnectOandaInput:
      type: object
      properties:
        apiToken:
          type: string
          minLength: 1
        accountId:
          type: string
          minLength: 1
        mode:
          type: string
          enum: [practice, live]
      required: [apiToken, accountId, mode]
    OandaStatus:
      type: object
      properties:
        connected:
          type: boolean
        mode:
          type: string
          enum: [practice, live]
        accountIdMasked:
          type: string
        currency:
          type: string
        balance:
          type: number
        nav:
          type: number
        unrealizedPl:
          type: number
        marginAvailable:
          type: number
        openPositionCount:
          type: number
        marginUsed:
          type: number
        marginRate:
          type: number
        openTradeCount:
          type: number
        pendingOrderCount:
          type: number
        realizedPl:
          type: number
        withdrawalLimit:
          type: number
        alias:
          type: string
        createdTime:
          type: string
      required: [connected]
    ConnectKrakenInput:
      type: object
      properties:
        apiKey:
          type: string
          minLength: 1
        apiSecret:
          type: string
          minLength: 1
      required: [apiKey, apiSecret]
    KrakenStatus:
      type: object
      properties:
        connected:
          type: boolean
        mode:
          type: string
          enum: [live]
        apiKeyMasked:
          type: string
        cash:
          type: number
        equity:
          type: number
      required: [connected]
    KrakenAutoTradeSettings:
      type: object
      properties:
        enabled:
          type: boolean
        maxPerTradeUsd:
          type: number
        dailyLimitUsd:
          type: number
        spentTodayUsd:
          type: number
        mode:
          type: string
          enum: [live]
      required: [enabled, maxPerTradeUsd, dailyLimitUsd, spentTodayUsd, mode]
    OandaAutoTradeSettings:
      type: object
      properties:
        enabled:
          type: boolean
        maxPerTradeUsd:
          type: number
        dailyLimitUsd:
          type: number
        spentTodayUsd:
          type: number
        mode:
          type: string
          enum: [practice, live]
      required: [enabled, maxPerTradeUsd, dailyLimitUsd, spentTodayUsd, mode]
    BrokerOrder:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
        notional:
          type: number
          nullable: true
        qty:
          type: number
          nullable: true
        filledQty:
          type: number
          nullable: true
        filledAvgPrice:
          type: number
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required: [id, symbol, side, status, createdAt]
...
    Trade:
      type: object
      required:
        [
          id,
          symbol,
          name,
          side,
          quantity,
          price,
          total,
          createdAt,
          logoUrl,
          broker,
          status,
        ]
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        price:
          type: number
        total:
          type: number
        createdAt:
          type: string
        logoUrl:
          type: string
        broker:
          type: string
        status:
          type: string
    TradeInput:
      type: object
      required: [symbol, side, notional]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        notional:
          type: number
          minimum: 1
...
    AutoTradeSettings:
      type: object
      required:
        [id, enabled, riskLevel, maxTradeAmountUsd, intervalMinutes, lastRunAt]
      properties:
        id:
          type: integer
        enabled:
          type: boolean
        riskLevel:
          type: string
          enum: [conservative, moderate, aggressive]
        maxTradeAmountUsd:
          type: number
        intervalMinutes:
          type: integer
        runMode:
          type: string
          enum: [interval, scheduled, continuous]
        scheduledTime:
          type: string
          nullable: true
        assetClasses:
          type: string
        lastRunAt:
          type: string
          nullable: true
        exitGuardsEnabled:
          type: boolean
        takeProfitPct:
          type: number
          nullable: true
        stopLossPct:
          type: number
          nullable: true
        dayTradeEnabled:
          type: boolean
        dayTradeMaxUsd:
          type: number
    AutoTradeSettingsInput:
      type: object
      properties:
        enabled:
          type: boolean
        riskLevel:
          type: string
          enum: [conservative, moderate, aggressive]
        maxTradeAmountUsd:
          type: number
        intervalMinutes:
          type: integer
        runMode:
          type: string
          enum: [interval, scheduled, continuous]
        scheduledTime:
          type: string
          nullable: true
          pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$"
        assetClasses:
          type: string
          pattern: "^(crypto|stock|forex|futures|commodity)(,(crypto|stock|forex|futures|commodity)){0,4}$"
        exitGuardsEnabled:
          type: boolean
        takeProfitPct:
          type: number
          nullable: true
          minimum: 0.1
          maximum: 1000
        stopLossPct:
          type: number
          nullable: true
          minimum: 0.1
          maximum: 99
        dayTradeEnabled:
          type: boolean
        dayTradeMaxUsd:
          type: number
          minimum: 1
          maximum: 1000000
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    AuthCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        confirmPassword:
          type: string
        firstName:
          type: string
        lastName:
          type: string
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...